Wireframe Projects - Simple First-Person Shooter - Done.
This was taken from issue 3 of Wireframe.

It was simple and the tutorial didn't really teach that much but there was some syntax I hadn't seen before so I added that to my list.
There was a weird bug with the object not shooting but I eventually found the inn-editor error and fixed it.